eatgo-api/src/test/java/me/jjeda/restaruant/interfaces/RestaruantControllerTest.java
package me.jjeda.restaruant.interfaces;

import org.hamcrest.core.StringContains;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.junit.Assert.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@WebMvcTest(RestaruantController.class)
public class RestaruantControllerTest {

    @Autowired
    private MockMvc mvc;

    @Test
    public void list() throws Exception {
        mvc.perform(get("/restaruant"))
                .andExpect(status().isOk())
                .andExpect(content().string(StringContains.containsString("\"name\":\"Bob Zip\"")));
    }

    public void detail() throws Exception {
        mvc.perform(get("/restauant/1"))
                .andExpect(status().isOk());
    }

}